Modified code to be more clear
