+++ src/styles.css
@@ -3,8 +3,6 @@
 
 @custom-variant dark (&:is(.dark *));
 
-@tailwind base;
-@tailwind components;
 @layer utilities;
 
 :root {
